refactor(llc): clean up LowLevelClientWriter
- extract `ResponseClassiferType` into `Autorest.CSharp.Common.Output.Models.Responses` namespace
- move generating `ResponseClassiferType` from `LowLevelClientWriter` to `RestClient`
- extract static definition from `LowLevelClientWriter.WriteCancellationTokenToRequestContextMethod()`

part of #2552
